Rename to NotificationsContainer
